Move static site to npm module
