+++ cldoc-static/README.md
@@ -1,0 +1,3 @@
+# cldoc-static
+
+cldoc-static is a static site generator for cldoc (see https://github.com/jessevdk/cldoc).
